+++ client/rest-high-level/src/test/java/org/elasticsearch/client/RequestConvertersTests.java
@@ -55,13 +55,6 @@ import org.elasticsearch.action.support.master.MasterNodeRequest;
 import org.elasticsearch.action.support.replication.ReplicationRequest;
 import org.elasticsearch.action.update.UpdateRequest;
 import org.elasticsearch.client.RequestConverters.EndpointBuilder;
-import org.elasticsearch.client.indexlifecycle.GetLifecyclePolicyRequest;
-import org.elasticsearch.client.indexlifecycle.LifecycleManagementStatusRequest;
-import org.elasticsearch.client.indexlifecycle.LifecyclePolicy;
-import org.elasticsearch.client.indexlifecycle.PutLifecyclePolicyRequest;
-import org.elasticsearch.client.indexlifecycle.DeleteLifecyclePolicyRequest;
-import org.elasticsearch.client.indexlifecycle.RetryLifecyclePolicyRequest;
-import org.elasticsearch.client.indexlifecycle.RemoveIndexLifecyclePolicyRequest;
 import org.elasticsearch.common.CheckedBiConsumer;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.bytes.BytesArray;
@@ -88,10 +81,6 @@ import org.elasticsearch.index.reindex.DeleteByQueryRequest;
 import org.elasticsearch.index.reindex.ReindexRequest;
 import org.elasticsearch.index.reindex.RemoteInfo;
 import org.elasticsearch.index.reindex.UpdateByQueryRequest;
-import org.elasticsearch.client.indexlifecycle.ExplainLifecycleRequest;
-import org.elasticsearch.client.indexlifecycle.SetIndexLifecyclePolicyRequest;
-import org.elasticsearch.client.indexlifecycle.StartILMRequest;
-import org.elasticsearch.client.indexlifecycle.StopILMRequest;
 import org.elasticsearch.rest.action.search.RestSearchAction;
 import org.elasticsearch.script.Script;
 import org.elasticsearch.script.ScriptType;
@@ -130,7 +119,6 @@ import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonMap;
 import static org.elasticsearch.client.RequestConverters.REQUEST_BODY_CONTENT_TYPE;
 import static org.elasticsearch.client.RequestConverters.enforceSameContentType;
-import static org.elasticsearch.client.indexlifecycle.LifecyclePolicyTests.createRandomPolicy;
 import static org.elasticsearch.index.query.QueryBuilders.matchAllQuery;
 import static org.elasticsearch.search.RandomSearchRequestGenerator.randomSearchRequest;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertToXContentEquivalent;
@@ -1469,143 +1457,6 @@ public class RequestConvertersTests extends ESTestCase {
                 + "previous requests have content-type [" + xContentType + "]", exception.getMessage());
     }
 
-    public void testGetLifecyclePolicy() {
-        String[] policies = rarely() ? null : randomIndicesNames(0, 10);
-        GetLifecyclePolicyRequest req = new GetLifecyclePolicyRequest(policies);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.getLifecyclePolicy(req);
-        assertEquals(request.getMethod(), HttpGet.METHOD_NAME);
-        String policiesStr = Strings.arrayToCommaDelimitedString(policies);
-        assertEquals(request.getEndpoint(), "/_ilm" + (policiesStr.isEmpty() ? "" : ("/" + policiesStr)));
-        assertEquals(request.getParameters(), expectedParams);
-    }
-
-    public void testPutLifecyclePolicy() throws Exception {
-        String name = randomAlphaOfLengthBetween(2, 20);
-        LifecyclePolicy policy = createRandomPolicy(name);
-        PutLifecyclePolicyRequest req = new PutLifecyclePolicyRequest(policy);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.putLifecyclePolicy(req);
-        assertEquals(HttpPut.METHOD_NAME, request.getMethod());
-        assertEquals("/_ilm/" + name, request.getEndpoint());
-        assertEquals(expectedParams, request.getParameters());
-    }
-
-    public void testDeleteLifecycle() {
-        String lifecycleName = randomAlphaOfLengthBetween(2,20);
-        DeleteLifecyclePolicyRequest req = new DeleteLifecyclePolicyRequest(lifecycleName);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.deleteLifecyclePolicy(req);
-        assertEquals(request.getMethod(), HttpDelete.METHOD_NAME);
-        assertEquals(request.getEndpoint(), "/_ilm/" + lifecycleName);
-        assertEquals(request.getParameters(), expectedParams);
-    }
-
-    public void testSetIndexLifecyclePolicy() throws Exception {
-        SetIndexLifecyclePolicyRequest req = new SetIndexLifecyclePolicyRequest();
-        String policyName = randomAlphaOfLength(10);
-        String[] indices = rarely() ? null : randomIndicesNames(0, 10);
-        req.policy(policyName);
-        req.indices(indices);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req, expectedParams);
-        setRandomIndicesOptions(req::indicesOptions, req::indicesOptions, expectedParams);
-
-        Request request = RequestConverters.setIndexLifecyclePolicy(req);
-        assertThat(request.getMethod(), equalTo(HttpPut.METHOD_NAME));
-        String idxString = Strings.arrayToCommaDelimitedString(indices);
-        assertThat(request.getEndpoint(),
-            equalTo("/" + (idxString.isEmpty() ? "" : (idxString + "/")) +
-                "_ilm/" + policyName));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testRemoveIndexLifecyclePolicy() {
-        Map<String, String> expectedParams = new HashMap<>();
-        String[] indices = randomIndicesNames(0, 10);
-        IndicesOptions indicesOptions = setRandomIndicesOptions(IndicesOptions.strictExpandOpen(), expectedParams);
-        RemoveIndexLifecyclePolicyRequest req = new RemoveIndexLifecyclePolicyRequest(Arrays.asList(indices), indicesOptions);
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.removeIndexLifecyclePolicy(req);
-        assertThat(request.getMethod(), equalTo(HttpDelete.METHOD_NAME));
-        String idxString = Strings.arrayToCommaDelimitedString(indices);
-        assertThat(request.getEndpoint(), equalTo("/" + (idxString.isEmpty() ? "" : (idxString + "/")) + "_ilm"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testStartILM() throws Exception {
-        StartILMRequest req = new StartILMRequest();
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.startILM(req);
-        assertThat(request.getMethod(), equalTo(HttpPost.METHOD_NAME));
-        assertThat(request.getEndpoint(), equalTo("/_ilm/start"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testStopILM() throws Exception {
-        StopILMRequest req = new StopILMRequest();
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.stopILM(req);
-        assertThat(request.getMethod(), equalTo(HttpPost.METHOD_NAME));
-        assertThat(request.getEndpoint(), equalTo("/_ilm/stop"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testLifecycleManagementStatus() throws Exception {
-        LifecycleManagementStatusRequest req = new LifecycleManagementStatusRequest();
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-
-        Request request = RequestConverters.lifecycleManagementStatus(req);
-        assertThat(request.getMethod(), equalTo(HttpGet.METHOD_NAME));
-        assertThat(request.getEndpoint(), equalTo("/_ilm/status"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testExplainLifecycle() throws Exception {
-        ExplainLifecycleRequest req = new ExplainLifecycleRequest();
-        String[] indices = rarely() ? null : randomIndicesNames(0, 10);
-        req.indices(indices);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req, expectedParams);
-        setRandomIndicesOptions(req::indicesOptions, req::indicesOptions, expectedParams);
-
-        Request request = RequestConverters.explainLifecycle(req);
-        assertThat(request.getMethod(), equalTo(HttpGet.METHOD_NAME));
-        String idxString = Strings.arrayToCommaDelimitedString(indices);
-        assertThat(request.getEndpoint(), equalTo("/" + (idxString.isEmpty() ? "" : (idxString + "/")) + "_ilm/explain"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
-
-    public void testRetryLifecycle() throws Exception {
-        String[] indices = randomIndicesNames(1, 10);
-        RetryLifecyclePolicyRequest req = new RetryLifecyclePolicyRequest(indices);
-        Map<String, String> expectedParams = new HashMap<>();
-        setRandomMasterTimeout(req::setMasterTimeout, TimedRequest.DEFAULT_MASTER_NODE_TIMEOUT, expectedParams);
-        setRandomTimeoutTimeValue(req::setTimeout, TimedRequest.DEFAULT_ACK_TIMEOUT, expectedParams);
-        Request request = RequestConverters.retryLifecycle(req);
-        assertThat(request.getMethod(), equalTo(HttpPost.METHOD_NAME));
-        String idxString = Strings.arrayToCommaDelimitedString(indices);
-        assertThat(request.getEndpoint(), equalTo("/" + (idxString.isEmpty() ? "" : (idxString + "/")) + "_ilm/retry"));
-        assertThat(request.getParameters(), equalTo(expectedParams));
-    }
 
     /**
      * Randomize the {@link FetchSourceContext} request parameters.
@@ -1756,7 +1607,7 @@ public class RequestConvertersTests extends ESTestCase {
         }
     }
 
-    private static void setRandomTimeoutTimeValue(Consumer<TimeValue> setter, TimeValue defaultTimeout,
+    static void setRandomTimeoutTimeValue(Consumer<TimeValue> setter, TimeValue defaultTimeout,
                                                   Map<String, String> expectedParams) {
         if (randomBoolean()) {
             TimeValue timeout = TimeValue.parseTimeValue(randomTimeValue(), "random_timeout");
@@ -1787,7 +1638,7 @@ public class RequestConvertersTests extends ESTestCase {
         }
     }
 
-    private static void setRandomMasterTimeout(Consumer<TimeValue> setter, TimeValue defaultTimeout, Map<String, String> expectedParams) {
+    static void setRandomMasterTimeout(Consumer<TimeValue> setter, TimeValue defaultTimeout, Map<String, String> expectedParams) {
         if (randomBoolean()) {
             TimeValue masterTimeout = TimeValue.parseTimeValue(randomTimeValue(), "random_master_timeout");
             setter.accept(masterTimeout);
